add entry page
