+++ frontend/package.json
@@ -13,6 +13,7 @@
     "axios": "^1.7.7",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^7.1.1",
     "react-scripts": "5.0.1",
     "typescript": "^4.9.5",
     "uuid": "^11.0.3",
